+++ CHANGES.rst
@@ -10,6 +10,8 @@ Changelog
   and not only the `Owner`.  `Owner` is not more managed by `meeting_workflow`.
   Fix meeting `review_state` could not be changed by another user than the `Owner`.
   [gbastien]
+- Fixed publication effective date not reindexed after publish.
+  [gbastien]
 
 2.0.0 (2024-09-23)
 ------------------